+++ src/AuditTrail/Abstractions/IAuditTrailService.cs
@@ -1,12 +1,11 @@
 ﻿using AuditTrail.Models;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 
 namespace AuditTrail.Abstractions;
 public interface IAuditTrailService<TPermission>
 {
-    Task SendToConsumerAsync(CancellationToken cancellationToken = default);
+    Task SendToTransactionConsumerAsync(TransactionEndEventData eventData, CancellationToken cancellationToken = default);
     Task<IEnumerable<AuditTrailDataBeforeSave<TPermission>>> GetEntityTrackedPropertiesBeforeSave(DbContextEventData eventData, CancellationToken cancellationToken = default);
     IEnumerable<AuditTrailDataAfterSave<TPermission>> UpdateEntityPropertiesAfterSave(IEnumerable<AuditTrailDataBeforeSave<TPermission>> auditEntitiesData,
         DbContext context);
@@ -20,4 +19,3 @@ public interface IAuditTrailService<TPermission, TInstance>
     : IAuditTrailService<TPermission>
 {
 }
-