+++ src/AuditTrail/Fluent/AuditTrailAssemblyProvider.cs
@@ -1,5 +1,4 @@
 ﻿using AuditTrail.Fluent.Abstractions;
-using System.Security;
 using static AuditTrail.Fluent.AssemblyScanner;
 
 namespace AuditTrail.Fluent;